+++ modules/security-center/resources.tf
@@ -1,6 +1,8 @@
 resource "azurerm_security_center_subscription_pricing" "pricing" {
+  for_each = toset(var.security_center_pricing_resource_types)
+
   tier          = var.security_center_pricing_tier
-  resource_type = var.security_center_pricing_resource_type
+  resource_type = each.value
 }
 
 resource "azurerm_security_center_contact" "contact" {